Move role enum to ~/common
