+++ src/components/authorization/dto/role.dto.ts
@@ -1,38 +1,6 @@
-import { ObjectType, registerEnumType } from '@nestjs/graphql';
-import { SecuredEnumList } from '~/common';
-import { ResourcesGranter } from '../policy';
+import { Role } from '~/common';
 
-export enum Role {
-  Administrator = 'Administrator',
-  BetaTester = 'BetaTester',
-  BibleTranslationLiaison = 'BibleTranslationLiaison',
-  Consultant = 'Consultant',
-  ConsultantManager = 'ConsultantManager',
-  Controller = 'Controller',
-  ExperienceOperations = 'ExperienceOperations',
-  FieldOperationsDirector = 'FieldOperationsDirector',
-  FieldPartner = 'FieldPartner',
-  FinancialAnalyst = 'FinancialAnalyst',
-  Fundraising = 'Fundraising',
-  Intern = 'Intern',
-  LeadFinancialAnalyst = 'LeadFinancialAnalyst',
-  Leadership = 'Leadership',
-  Liaison = 'Liaison',
-  Marketing = 'Marketing',
-  Mentor = 'Mentor',
-  ProjectManager = 'ProjectManager',
-  RegionalCommunicationsCoordinator = 'RegionalCommunicationsCoordinator',
-  RegionalDirector = 'RegionalDirector',
-  StaffMember = 'StaffMember',
-  Translator = 'Translator',
-}
-
-registerEnumType(Role, { name: 'Role' });
-
-@ObjectType({
-  description: SecuredEnumList.descriptionFor('roles'),
-})
-export abstract class SecuredRoles extends SecuredEnumList<Role, Role>(Role) {}
+export * from '~/common/role.dto';
 
 export type ProjectScope = 'project';
 export type GlobalScope = 'global';
@@ -55,17 +23,3 @@ export const withoutScope = (role: ScopedRole): Role => splitScope(role)[1];
 
 export const splitScope = (role: ScopedRole) =>
   role.split(':') as [AuthScope, Role];
-
-// eslint-disable-next-line @typescript-eslint/no-namespace
-export namespace Role {
-  /**
-   * A helper to grant roles to be assignable in a more readable way.
-   * This should be used within a Policy.
-   */
-  export const assignable = (
-    resources: ResourcesGranter,
-    roles: readonly Role[]
-  ) => resources.AssignableRoles.grant(roles);
-
-  Object.defineProperty(Role, 'assignable', { enumerable: false });
-}
